requirements.txt
# 核心套件
google-genai>=0.2.0
python-dotenv>=1.0.0

# 爬蟲相關
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
trafilatura>=1.6.0
googlesearch-python>=1.2.0

# 數據處理
pandas>=2.0.0
numpy>=1.24.0
jieba>=0.42.1

# 視覺化（選裝）
matplotlib>=3.7.0

# 報告生成（選裝）
markdown>=3.5.0
pdfkit>=1.0.0

# 工具
tqdm>=4.66.0
